Merge r5693:5712 from /nmap-exp/david/nmap-massping-migration.
Always update srtt, rttvar, and timeout for every response, even if we don't adjust congestion control or send delay variables.

Be more careful about checking gstats->sendOK when sending retransmits.
Previously, it was only checked once per traversal of the incomplete
hosts list, which meant that enough probes could be sent in a round to
exceed the congestion window. Explanatory pictures are at
http://www.bamsoftware.com/wiki/Nmap/PerformanceGraphs#retransmit-sendOK.

This needs some more testing to see what effect it has on scan times. My
instinct says it will slow them down, because retransmits will be sent
no faster than before, and retransmits will be more likely to be
responded to, leading to more drops. On the other hand, correctly
detecting a drop and marking a host up is better than blasting
retransmits faster than they can be responded to.



git-svn-id: svn://svn.insecure.org/nmap@5713 e0a8ed71-7df4-0310-8962-fdc924857419