Disable auto-generation of libpcre/pcre_chartables.c. That file is meant to be automatically generated by the makefile; however we include a copy to enable the Windows build to work. On Unix, the makefile was overwriting the file with a symbolic link to libpcre/pcre_chartables.c.dist, which was causing revision-control headaches. Now that file is copied instead of linked to, so its contents don't change.
It is safe to use a static copy of the file instead of an auto-generated
one because Nmap has never used the auto-generation feature. You need to
use a special flag --enable-rebuild-chartables to get that. So copying
the file has the same behavior as creating a link. Furthermore, it's not
desirable to generate that file (using dftables, an auxiliary program
that uses the current locale) because we don't want our regular
expressions to be locale-dependent.

pcre_chartables.c was also removed from CLEANFILES so it's not deleted
by "make clean" or "make distclean". pcre.h is still deleted by "make
distclean" because it's included in AC_CONFIG_FILES. That's not as big a
deal.


git-svn-id: svn://svn.insecure.org/nmap@5927 e0a8ed71-7df4-0310-8962-fdc924857419