Merge from /nmap-exp/david/nmap-closed-ipid.
This adds a new SEQ.CI OS detection test. It is based on the IP IDs of the
responses to the three probes sent to a closed port: T5, T6, and T7. This has
been reported to differ from TI and was observed to do so in Internet scans
using the nmap-closed-ipid branch.


git-svn-id: svn://svn.insecure.org/nmap@12292 e0a8ed71-7df4-0310-8962-fdc924857419