Merge from /nmap-exp/david/nselib-http. This is an overhaul of HTTP parsing mostly. Response parsing is centralized, and fewer operations are done on raw HTTP data.
The biggest user-visible change is that http.request goes back to
returning a parsed result table, not raw HTTP data. I believe this is
how the function worked in the past; it's what the NSEDoc for the
function says. The only thing that used http.request was citrixxml.lua,
and this commit alters it to match the new expectations.

The other change is that the http.pipeline function no longer accepts
the "raw" option. The only script that used that was sql-injection.nse,
and this commit modifies that script as well.


git-svn-id: svn://svn.insecure.org/nmap@16447 e0a8ed71-7df4-0310-8962-fdc924857419