Don't pass the cookies table as a parameter, but as a member of the options table, in http functions. It was unreasonable that this yet-unused feature was given a more prominent place than even the header and request body, both of which are in the options table.
This change doesn't affect any other scripts or libraries because none
of them use cookies. In the cases, like http.get, where cookies was an
optional final parameter, I just removed it. Where it was not the final
parameter, as in http.post and http.pGet, I left the parameter in place
but documented that it is ignored for backwards compatibility.


git-svn-id: svn://svn.insecure.org/nmap@16823 e0a8ed71-7df4-0310-8962-fdc924857419