Accuracy and speed improvements to irc-unrealircd-backdoor.nse. The main problem was that the timer was including the time taken by the server to do reverse-DNS resolution and an ident lookup, before it handled the magic "AB" sleep command. So if a server took 10 seconds to time out the ident lookup, the time recorded would be 10 seconds greater than it should have been. If the timeout was long enough, as server would be reported as vulnerable even if not. To compensate for this, the delay was set high, but false positives were still possible and this slowed down the script.
First, remove the mutex that allowed only one instance of the script to
run at a time. The mutex was meant to provide more accurate timing, but
it wasn't really needed because scritp parallelism wasn't the cause of
inaccuracy.

Next, make sure the server is done sending its initialization messages
(and hence is done with its timeouts) before sending the magic "AB"
sleep command and starting the timer. This is done by sending an
innocuous TIME command immediately upon connection. This is partly
because comm.tryssl has to send something, and partly because we can
detect when the server has processed the TIME command by looking for a
message like ":hostname 451 TIME :You have not registered". Once this is
done, we start the timer and send the "AB" command. If we haven't gotten
a response to TIME in 60 seconds, go ahead and send the "AB" command
anyway.

Finally, reduce the delay to 8 seconds from 25. The delay only has to be
long enough to overcome any network delay, now that confounding timeouts
have been eliminated. In a test of around 600 hosts, the only times I
recorded were 0, 1, 8, and 9 seconds, so this looks like a good enough
safety margin.


git-svn-id: svn://svn.insecure.org/nmap@18508 e0a8ed71-7df4-0310-8962-fdc924857419