Change the warning message for time specifications that look off by a factor of 1000. I realized there is a bad ambiguity in
The default unit for --host-timeout is seconds (since April 2010),
so your time of "10000" is 2.8 hours.

It could be interpreted as "seconds since April 2010" which is a strange
way to specify a length of time and would be confusing if you took it
that way. I have changed it to

Since April 2010, the default unit for --host-timeout is seconds,
so your time of "10000" is 2.8 hours.


git-svn-id: svn://svn.insecure.org/nmap@18791 e0a8ed71-7df4-0310-8962-fdc924857419